server: fix: move cors middleware to the separate middleware file
